- dont override operator new in linux - squashed some warnings

+++ src/application/gameobject.cpp
@@ -55,7 +55,7 @@ namespace _462
 
 	GameObject::~GameObject()
 	{
-		for (int i = 0; i < components.size(); i++)
+		for (size_t i = 0; i < components.size(); i++)
 		{
 			delete components[i];
 		}
